+++ src/server-api.mjs
@@ -62,9 +62,10 @@ export const deleteContactsOnServer = async (clientId) => {
     }
 }
 
-export const addContactToKanbanOnServer = async (contactId) => {
+export const addContactToKanbanOnServer = async (contactId, type) => {
+
     try {
-        const response = await fetch('/api/new-task', {
+        const response = await fetch(`/api/new-task/${type}`, {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
@@ -83,10 +84,10 @@ export const addContactToKanbanOnServer = async (contactId) => {
     }
 }
 
-export const deleteContactFromKanbanOnServer = async (clientId) => {
+export const deleteContactFromKanbanOnServer = async (clientId, type) => {
     try {
         console.log('@CLIENT_ID to delete from KANBAN:', clientId)
-        const response = await fetch(`/api/delete-task/${clientId}`, {
+        const response = await fetch(`/api/delete-task/${type}/${clientId}`, {
             method: 'DELETE',
             headers: {
                 'Content-Type': 'application/json',
@@ -104,10 +105,10 @@ export const deleteContactFromKanbanOnServer = async (clientId) => {
     }
 }
 
-export const updateTaskPositionOnServer = async (taskId, newPosition) => {
+export const updateTaskPositionOnServer = async (taskId, newPosition, type) => {
     try {
         const response = await fetch(
-            `/api/update-task-position/${taskId}/${newPosition}`, {
+            `/api/update-task-position/${type}/${taskId}/${newPosition}`, {
             method: 'PUT',
             headers: {
                 'Content-Type': 'application/json',
@@ -125,14 +126,14 @@ export const updateTaskPositionOnServer = async (taskId, newPosition) => {
     }
 }
 
-export const updateTaskDetailsOnServer = async (taskId, updatedDetails) => {
+export const updateTaskDetailsOnServer = async (taskId, updatedDetails, type) => {
     try {
         const response = await fetch(
-            `/api/update-task-details/${taskId}`, {
+            `/api/update-task-details/${type}/${taskId}`, {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
-            },
+            },  
             body: JSON.stringify(updatedDetails),
         })